+++ tsup.config.ts
@@ -25,7 +25,6 @@ const CLIENT_EXTERNAL = [
   'clsx',
   'tailwind-merge',
   /^@radix-ui\/.+/,
-  /^@abeta\.dev\/auth(\/.+)?$/,
 ];
 
 export default defineConfig([